make it better

+++ src/main/java/com/springproject/blogapplication/controller/CommentController.java
@@ -1,12 +1,16 @@
 package com.springproject.blogapplication.controller;
 
 import com.springproject.blogapplication.model.Comment;
+import com.springproject.blogapplication.model.Post;
 import com.springproject.blogapplication.model.Tag;
+import com.springproject.blogapplication.model.User;
 import com.springproject.blogapplication.service.CommentService;
 import com.springproject.blogapplication.service.PostService;
 import com.springproject.blogapplication.service.TagService;
 
+import com.springproject.blogapplication.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -16,10 +20,15 @@ public class CommentController {
 
     @Autowired
     private CommentService commentService;
+
     @Autowired
     private PostService postService;
+
     @Autowired
     private TagService tagService;
+
+    @Autowired
+    private UserService userService;
 
     @GetMapping("/save/{id}")
     public String commentForm(@PathVariable("id") Integer id,
@@ -31,19 +40,32 @@ public class CommentController {
     }
 
     @PostMapping("/user/deleteComment/{commentId}")
-    public String deleteComment(@PathVariable("commentId") Integer commentId) {
+    public String deleteComment(@PathVariable("commentId") Integer commentId, Authentication authentication) {
         Integer postId = commentService.getPostIdFromComments(commentId);
-        commentService.deleteCommentById(commentId);
+        User user = userService.getUserByEmail(authentication.getName());
 
+        Post post=postService.getPostById(postId);
+
+        if(post.getAuthor().equals(user.getUserName())) {
+
+            commentService.deleteCommentById(commentId);
+        }
         return "redirect:/blog/" + postId;
     }
 
     @GetMapping("/user/updateCommentForm/{commentId}")
-    public String updateCommentForm(@PathVariable("commentId") Integer commentId, Model model) {
+    public String updateCommentForm(@PathVariable("commentId") Integer commentId, Model model,  Authentication authentication) {
         Comment comment = commentService.getCommentById(commentId);
-        model.addAttribute("comment", comment);
+        Integer postId = commentService.getPostIdFromComments(commentId);
+        Post post=postService.getPostById(postId);
+        User user = userService.getUserByEmail(authentication.getName());
 
-        return "commentForm";
+        if(post.getAuthor().equals(user.getUserName())) {
+
+            model.addAttribute("comment", comment);
+            return "commentForm";
+        }
+        return "redirect:/blog/"+postId;
     }
 
     @PostMapping("/saveUpdate/{commentId}")
